Added XMLDoc and checks against null.

+++ FlorianWolters.Office.Word.AddIn.CBA/Commands/RefreshCustomXMLPartsCommand.cs
@@ -13,8 +13,16 @@ namespace FlorianWolters.Office.Word.AddIn.CBA.Commands
     using FlorianWolters.Office.Word.Commands;
     using Word = Microsoft.Office.Interop.Word;
 
+    /// <summary>
+    /// The <i>Command</i> <see cref="RefreshCustomXMLPartsCommand"/> refreshes
+    /// all custom XML parts in the active Microsoft Word document with the
+    /// content from the XML files in a subdirectory of that document.
+    /// </summary>
     internal class RefreshCustomXMLPartsCommand : ApplicationCommand
     {
+        /// <summary>
+        /// The name of the directory with the XML files.
+        /// </summary>
         private readonly string xmlDirectoryName;
 
         /// <summary>
@@ -37,16 +45,25 @@ namespace FlorianWolters.Office.Word.AddIn.CBA.Commands
         {
             Word.Document document = this.Application.ActiveDocument;
 
-            CustomXMLPartRepository customXMLPartRepository = new CustomXMLPartRepository(document.CustomXMLParts);
+            if (null != document)
+            {
+                CustomXMLPartRepository customXMLPartRepository = new CustomXMLPartRepository(document.CustomXMLParts);
 
-            // The approach is very simple:
-            // 1. Delete all custom XML parts which are user defined.
-            // 2. Add all XML files in the specified subdirectory as new custom XML parts.
-            customXMLPartRepository.DeleteAllNotBuiltIn();
-            string directoryPath = this.DirectoryPathForXMLFiles(document);
-            customXMLPartRepository.AddFromDirectory(directoryPath);
+                // The approach is very simple:
+                // 1. Delete all custom XML parts which are user defined.
+                // 2. Add all XML files in the specified subdirectory as new custom XML parts.
+                customXMLPartRepository.DeleteAllNotBuiltIn();
+                string directoryPath = this.DirectoryPathForXMLFiles(document);
+                customXMLPartRepository.AddFromDirectory(directoryPath);
+            }
         }
 
+        /// <summary>
+        /// Retrieves the absolute directory path of the directory with the XML
+        /// files for the specified <see cref="Word.Document"/>.
+        /// </summary>
+        /// <param name="document">The <see cref="Word.Document"/> whose XML directory path to retrieve.</param>
+        /// <returns>The absolute directory path of the XML directory for the specified <see cref="Word.Document"/>.</returns>
         private string DirectoryPathForXMLFiles(Word.Document document)
         {
             return document.Path + Path.DirectorySeparatorChar + this.xmlDirectoryName;
